Add importer for Passman
The importer reads JSON files generated by the Nextcloud app Passman.

The first tag is used as the folder name. This works well if passwords
have no or only one tag. If no username is set then the email address is
used as the username.

Files are not being imported.