Add Microsoft RDP URI schemes (#180)
This will allow users to launch the Remote Desktop client directly from
Bitwarden if their OS supports the URI scheme.

Per
https://docs.microsoft.com/en-us/windows-server/remote/remote-desktop-services/clients/remote-desktop-uri,
rdp:// is legacy and ms-rd: is the preferred one. I'm adding both so
that either can be used as ms-rd: is not available on all versions of
Windows and rdp:// is easier to add as a custom URI scheme in Windows 10
if it is desired.