Select an image to display for credit cards based on the brand.
The image will be any image with the same name as the brand + .png (e.g. "Mastercard.png"), located in the "cards" subfolder of "images" folder in the client.
If there is no image for the brand of the card, the generic "Other.png" image in the "cards" subfolder is chosen.

If the user has deactivated the display of logos, the display will be with font awesome icon, same as before this commit.wq