[bug] Return the correct environmentUrls when scaffloging a new account
To allow for setting environmentUrls before an account is created we save that value as a global setting and then apply it to any newly authed accounts.
There is a bug that will instead save the urls used by the previous logged in account, making account switching with multiple servers cause errors.

This commit resolves this by specifically getting environementUrls from global state when creating a new account