[bug] Set all urls when setting urls from storage
Previously when only one account could be authenticated at a time if a baseUrl was found we didn't need to worry about setting other urls in the environmentService, and simply returning

With account switching it is possible to have multiple url configurations, some with a baseUrl and some without, and this early return was causing accounts configured this way to clash. It became possible to have something like:

urls.baseUrl = 'https://testdo.bitwarden.com' set as the base url for an account that was intended to target the cloud, causing deauthenticated sessions.

This commit ensures that all urls are set each time getUrlsFromStorage is called.