[bug] Null check lists when changing the vault filter
Sometimes when changing an organization fault filter the system can end up in a postion where the filter modifier selected isn't valid for the newly selected organization.
The clearest example of this are collections, since changing any org filter with a collection already selected will make that collection selection invalid.

Because of this, when an organization filter is selected we check to make sure modifiers are still valid. If they aren't we clear them.
This isn't working if a list is checked and this list is null. An error throws and the filter doesn't change. This commit adds a null check to keep things moving.