[bug] Add missing null check in vault filters
Sometimes changing an organization filter involves resetting the filter modifier because the current one is not relevant to the new organization filter.
One example of this is if a collection is selected, as changing an organization filter will always invalidate a selected collection.
We are missing a null check when we check this to prune the filter, resulting in filter changes sometimes not being applied and console errors.

This commit adds a null check when correcting a requested filter change.